Migrates private AWS project data functions from JS API file to their own
